Extract ExpenseCard vom ExpenseList
